massive and suspicious performance boost in phase1 code: Fix A: epsilon 1e-4f → 1e-2f (escape self-intersection) Fix B: stable quadratic in Sphere::Intersect (no 1e10 - 1e10 cancellation) Fix C: normal computed from bounded quantities inside Intersect (no ~1e5 - ~1e5 in Normalize(p - m_p))

+++ phase1/src/geometry.cuh
@@ -22,11 +22,11 @@ namespace smallpt {
 		// Constructors and Destructors
 		//---------------------------------------------------------------------
 
-		__device__ explicit Ray(Vector3 o, 
-								Vector3 d, 
-								double tmin = 0.0, 
-								double tmax = std::numeric_limits< double >::infinity(), 
-								std::uint32_t depth = 0u) noexcept
+		__device__ explicit Ray(Vector3 o,
+		                        Vector3 d,
+		                        float tmin = 0.0f,
+		                        float tmax = std::numeric_limits< float >::infinity(),
+		                        std::uint32_t depth = 0u) noexcept
 			: m_o(std::move(o)),
 			m_d(std::move(d)),
 			m_tmin(tmin),
@@ -47,7 +47,7 @@ namespace smallpt {
 		// Member Methods
 		//---------------------------------------------------------------------
 
-		__device__ const Vector3 operator()(double t) const {
+		__device__ const Vector3 operator()(float t) const {
 			return m_o + m_d * t;
 		}
 
@@ -56,7 +56,7 @@ namespace smallpt {
 		//---------------------------------------------------------------------
 
 		Vector3 m_o, m_d;
-		mutable double m_tmin, m_tmax;
+		mutable float m_tmin, m_tmax;
 		std::uint32_t m_depth;
 	};
 }
